+++ test/SwaggerUiControllerFactoryTest.php
@@ -1,16 +1,15 @@
 <?php
 /**
  * @license   http://opensource.org/licenses/BSD-3-Clause BSD-3-Clause
- * @copyright Copyright (c) 2014-2016 Zend Technologies USA Inc. (http://www.zend.com)
+ * @copyright Copyright (c) 2014-2018 Zend Technologies USA Inc. (http://www.zend.com)
  */
 
 namespace ZFTest\Apigility\Documentation\Swagger;
 
-use PHPUnit_Framework_TestCase as TestCase;
+use PHPUnit\Framework\TestCase;
 use Zend\ModuleManager\ModuleManager;
 use Zend\ServiceManager\Exception\ServiceNotCreatedException;
 use Zend\ServiceManager\ServiceManager;
-use Zend\Test\PHPUnit\Controller\AbstractHttpControllerTestCase;
 use ZF\Apigility\Documentation\ApiFactory;
 use ZF\Apigility\Documentation\Swagger\SwaggerUiController;
 use ZF\Apigility\Documentation\Swagger\SwaggerUiControllerFactory;
@@ -35,10 +34,13 @@ class SwaggerUiControllerFactoryTest extends TestCase
         $this->services = $services = new ServiceManager();
     }
 
+    /**
+     * @expectedException \Zend\ServiceManager\Exception\ServiceNotCreatedException
+     */
     public function testExceptionThrownOnMissingApiCreatorClass()
     {
         $smFactory = $this->factory;
-        $this->setExpectedException(ServiceNotCreatedException::class);
+        $this->expectException(ServiceNotCreatedException::class);
         $factory = $smFactory($this->services, SwaggerUiController::class);
     }